+++ lib/wpt-tests/defineProperty.window.test.ts
@@ -19,7 +19,7 @@
 
 //#region Imports
 import { describe, expect, test } from "vitest";
-import { BrowserStorage } from "../index";
+import { BrowserStorage } from "../storage";
 //#endregion
 
 describe.sequential.each([9, "x"])("%j", function (key) {
