+++ lib/wpt-tests/missing_arguments.window.test.ts
@@ -19,7 +19,7 @@
 
 //#region Imports
 import { expect, test } from "vitest";
-import { BrowserStorage } from "../index";
+import { BrowserStorage } from "../storage";
 //#endregion
 
 test.each<(storage: Storage) => unknown>([
