+++ lib/wpt-tests/storage_getitem.window.test.ts
@@ -19,7 +19,7 @@
 
 //#region Imports
 import { describe, expect, test } from "vitest";
-import { BrowserStorage } from "../index";
+import { BrowserStorage } from "../storage";
 //#endregion
 
 describe("Get value by getItem(key) and named access", () => {
