+++ lib/wpt-tests/storage_indexing.window.test.ts
@@ -19,7 +19,7 @@
 
 //#region Imports
 import { describe, expect, test } from "vitest";
-import { BrowserStorage } from "../index";
+import { BrowserStorage } from "../storage";
 //#endregion
 
 describe("Indexed getter", function () {
