+++ lib/wpt-tests/storage_set_value_enumerate.window.test.ts
@@ -19,7 +19,7 @@
 
 //#region Imports
 import { expect, test } from "vitest";
-import { BrowserStorage } from "../index";
+import { BrowserStorage } from "../storage";
 //#endregion
 
 const store_list = [
