+++ lib/wpt-tests/storage_supported_property_names.window.test.ts
@@ -19,7 +19,7 @@
 
 //#region Imports
 import { expect, test } from "vitest";
-import { BrowserStorage } from "../index";
+import { BrowserStorage } from "../storage";
 //#endregion
 
 test("Object.getOwnPropertyNames on storage Storage", () => {
